add addon controller
